feat: added loyalty points and coupons in subscription manual order
