+++ src/apps/carts/components/index.js
@@ -1,5 +1,7 @@
 import Modifiers from './Modifiers'
 import ProductOptions from './ProductOptions'
 import QuantitySelector from './QuantitySelector'
+import LoyaltyPoints from './LoyalityPoints'
+import Coupon from './Coupons'
 
-export { Modifiers, ProductOptions, QuantitySelector }
+export { Modifiers, ProductOptions, QuantitySelector, LoyaltyPoints, Coupon }
