+++ src/apps/carts/views/forms/subscription/tunnels/index.js
@@ -1,2 +1,3 @@
 export * from './Fulfillment'
 export * from './Payment'
+export * from './Coupons'
